Check Homebrew location on M1
Before, the way the script was determining whether or not Homebrew was installed was by looking for the "brew" command.

If Homebrew was installed on an M1 Mac in Rosetta mode, the brew command will exist even when in native mode. However, Homebrew uses separate installation directories for Rosetta vs native, and my script assumed that because the brew command was found, Homebrew was installed in the right place, which was a wrong assumption to make.

Now, the script checks for the /opt/homebrew directory when in native mode, and if it doesn't find it, it installs Homebrew.